removed obsolete tx recovery facility, fixed bug handling non-eip1559 nodes
